// src/types/jellyfin.ts
export interface MediaItem {
  Name: string;
  Id: string;
  Type: string;
  ImageTags?: ImageTags;
  BackdropImageTags?: string[];
  ParentBackdropImageTags?: string[];
  ParentPrimaryImageTag?: string;
  ParentBackdropItemId?: string;
  ParentId?: string;
  SeriesName?: string;
  ProductionYear?: number;
  IndexNumber?: number;
  ParentIndexNumber?: number;
  PremiereDate?: string;
  Overview?: string;
  CommunityRating?: number;
  OfficialRating?: string;
  RunTimeTicks?: number;
  MediaType?: string;
  GenreItems?: Array<{ Name: string; Id: string }>;
  SeriesId?: string;
  SeasonId?: string;
  MediaStreams?: MediaStream[];
  Genres?: string[];
  CriticRating?: number;
  People?: Array<People>;
  Profession?: string;
  Studios?: Array<Studios>;
  RemoteTrailers?: Array<{
    Url: string;
  }>;
  UserData: {
    PlaybackPositionTicks?: number;
    Played?: boolean;
    IsFavorite?: boolean;
    IsRated?: boolean;
    PlayCount?: number;
    Key?: string;
    KeyType?: string;
    LastPlayedDate?: string;
    LastPlayedDateTicks?: number;
    LastPlayedDateString?: string;
  };
  ProductionLocations?: string[];
  ProviderIds?: ProviderIds;
  Chapters?: Array<{
    Name: string;
    StartPositionTicks: number;
    EndPositionTicks: number;
  }>;
}
export interface Studios {
  Name: string;
  Id: string;
}

type PeopleType = "Actor" | "Director" | "Producer" | "Writer";

export interface People {
  Name: string;
  Id: string;
  Role?: string;
  Type?: PeopleType;
  PrimaryImageTag?: string;
  ImageBlurHashes?: {
    Primary?: {
      [size: string]: string;
    };
  };
  SortName?: string;
  Path?: string;
  ProductionYear?: number;
  Overview?: string;
  IndexNumber?: number;
  ParentIndexNumber?: number;
  PremiereDate?: string;
  BirthDate?: string;
  DeathDate?: string;
  IsFavorite?: boolean;
  ExternalUrls?: {
    Name: string;
    Url: string;
  }[];
  ServerId?: string;
  ImageTags?: ImageTags;
  Profession?: string;
  ProductionLocations?: string[];
  ProviderIds?: ProviderIds;
}

export interface ProviderIds {
  Imdb: string;
}

export interface ImageTags {
  Primary?: string;
  Thumb?: string;
  Logo?: string;
  Backdrop?: string;
}
// ...
export interface MediaStream {
  Index: number;
  Codec: string;
  CodecTag?: string;
  Language?: string;
  LanguageCode?: string;
  IsInterlaced?: boolean;
  IsDefault?: boolean;
  IsForced?: boolean;
  Type: "Video" | "Audio" | "Subtitle";
  AspectRatio?: string;
  Channels?: number;
  SampleRate?: number;
  BitRate?: number;
  BitDepth?: number;
  Width?: number;
  Height?: number;
  AverageFrameRate?: number;
  RealFrameRate?: number;
  Profile?: string;
  IsTextSubtitleStream?: string;
  Level?: number;
  IsExternal?: boolean;
  DeliveryMethod?: string;
  DisplayTitle?: string;
  DisplayLanguage?: string;
  NalLengthSize?: string;
  Title?: string;
}